Added define test to support a reduction of file tests
In previous code, an image was provided to support display
of that image to click on for seeing an enlarged image.
If though it is known that no image is going to exist, then
it is now possible to establish a constant that when set
to false, will cause the code to not try to check for the
existence of the file 'bigger_picture.jpg'.
The constant could be defined in includes/extra_configures file
or in the database:
<?php
define('PRODUCTS_IMAGE_LARGER_TEXT_FILE_IMAGE', false);

should be sufficient.