Incorporate zcJS and tighter check for loading javascript code.
Added detection and use of the ZC 1.5.3+ built in communicator
basically for AJAX.  Code still functions through original
method if installed on an older ZC system, but uses the
newer zcJS variable when made available.

Also modified the onload function to skip the cart_quantity
"text" box because it is not likely to have an onchange
event assigned to it at the time of loading. But will likely
be listened to when changed in case there is something that
is supposed to happen differently based on mutiples.