Handle elements with no events on load
The on_load_ais.js file calls the javascript function ais_init(), which if the conditions support swapping images would also attempt to call applicable action items for each element.  The problem with this is that if an element does not have the action assigned, then it is possible that the browser will provide a javascript error.  This commit adds functionality to try the action and catch the error that occurs if there is one.  This supports continued operations in absence of some sort of complete rewrite of the plugin and specifically supports swapping out the image when returning to the product with attributes pre-selected (ie. from the shopping cart).