Add get_attrib_image function to observer
Added a help/utility function to simplify the coding of pulling an
attribute image from the database.  The function does not limit the
selections to just those that are swappable, but pulls any attribute
image from the products_attributes table.  This is further language
supportive in that it will pull the language associated with the current
browser setting, but if an image does not exist for that language it will
provide the image/image path for the default language.