Ensure variables defined.
Fixes #8

Address  LARGE_IMAGE_WIDTH and LARGE_IMAGE_HEIGHT not being defined

Have added in a check supporting php 7.x to use an empty text string if
the constants are not defined.

Address PRODUCTS_OPTIONS_TYPE_LINK not being defined.

The plugin does not require PRODUCTS_OPTIONS_TYPE_LINK to be defined in
order for it to work or to be used.  Also because the associated code is
in line with the normal catalog files, if the constant is not defined then
notices will get thrown and later versions of php there may be an error.