Improve zen_colorbox Incorporation.
Fixes #7

Added the loading of the zen_colorbox module file if
it exists.  Also converted the old style connection
from using the html tag rel= to data-cbox-rel as
used in the latest version of zen colorbox. Note that
by working with/through the file that the details of
this change are not important as they are/should be
handled by the zen colorbox code.