Add specific ZC 1.5.7 files
Although this updates the plugin to present the ZC 1.5.7 files,
it does not immediately ensure operation if upgrade/reinstallation
is as suggested in the forum.

The needs of this plugin can be met without modification of
other files even though there is one admin core file edit needed.
Ideally code could be incorporated into that file to omit the
need for a modification.

Because of changes to how the items are shown in ZC 1.5.7,
changes were needed for the module. To not lose the original
formatting, the individual versions are updated with their
specific language file.

Also this is instead of installing a "core" version of the
language file and then overwriting with a version specific
instance of the file.