Update SQL queries
Although no reports have been received in the related forum,
updated the temporary variable assignment to use a SELECT
query.

Also updated the INSERT queries to use IGNORE so that
needed results are obtained where not previously present and
no adverse action in the admin occurs. Unfortunately, doing
so does not provide an installer with significant information
about the database status in performing the installation.