Prevent continuous page reload
In Zen Cart 1.5.7, there were some additional sanitization
features added that would cause AJAX related calls that include
the URL parameter `products_id` to reload if not accompanied
by a `main_page` parameter. This adds the `main_page` parameter
if it is already on the path.

Why not remove the `products_id` parameter? The information
is beneficial in URL tracking associated with parameters.