Eliminate "hiding" script code
This was a practice incorporated into Zen Cart code for some
reason. It seemed necessary at a time, but now really just
makes it harder to review code changes because IDEs will
actually ignore these commented out code sections but browsers
will still execute the code. So it really doesn't serve
much of a purpose now-a-days.